algotrade-master/backend/pg_initer/requirements.txt
backoff==2.2.1
celery==5.3.4
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_settings==2.0.3